+++ cycle-analysis.Domain/SessionData/Dtos/SessionDataDto.cs
@@ -8,11 +8,11 @@
         public DateTime Date { get; set; }
         public DateTime Time { get; set; }
         public int Row { get; set; }
-        public int HeartRate { get; set; }
-        public int Speed { get; set; }
-        public int Cadence { get; set; }
-        public int Altitude { get; set; }
-        public int Power { get; set; }
+        public double HeartRate { get; set; }
+        public double Speed { get; set; }
+        public double Cadence { get; set; }
+        public double Altitude { get; set; }
+        public double Power { get; set; }
         public int SessionId { get; set; }
     }
 }
