+++ cycle-analysis.Domain/SessionData/Models/SessionData.cs
@@ -6,11 +6,11 @@ namespace cycle_analysis.Domain.SessionData.Models
     {
         public int Id { get; set; }
         public int Row { get; set; }
-        public int HeartRate { get; set; }
-        public int Speed { get; set; }
-        public int Cadence { get; set; }
-        public int Altitude { get; set; }
-        public int Power { get; set; }
+        public double HeartRate { get; set; }
+        public double Speed { get; set; }
+        public double Cadence { get; set; }
+        public double Altitude { get; set; }
+        public double Power { get; set; }
         public int SessionId { get; set; }
         public virtual Session Session { get; set; }
     }
